+++ R/crandb-public-api.R
@@ -161,6 +161,95 @@ cran_top_downloaded <- function() {
   tibble::as_tibble(tb)
 }
 
+#' New CRAN packages
+#'
+#' List the latest new CRAN packages.
+#'
+#' @param from Time interval to query. Possible values:
+#' * `"last-week"`
+#' * `"last-month"`
+#' * A [Date] object to be used as a start date.
+#' * A [POSIXt] object to be used as the start date.
+#' * A [difftime] object to used as the time interval until now.
+#' * An integer scalar, the number of days until today.
+#' * A character string that is converted to a start date using
+#'   [as.POSIXct()].
+#' @param last Integer to limit the number of returned packages.
+#' @return Tibble of package descriptions.
+#'
+#' @export
+#' @importFrom parsedate format_iso_8601
+#' @examples
+#' \dontrun{
+#' # Last week
+#' cran_new("last-week")
+#'
+#' # Last month
+#' cran_new("last-month")
+#'
+#' # Last 5 days
+#' cran_new(from = 5)
+#'
+#' # From a given date, but at most 10
+#' cran_new(from = "2021-04-06", last = 10)
+#' }
+
+cran_new <- function(from = "last-week", last = Inf) {
+
+  if (inherits(from, "Date") || inherits(from, "POSIXt")) {
+    from <- as.POSIXct(from)
+
+  } else if (inherits(from, "difftime")) {
+    from <- Sys.time() - from
+
+  } else if (identical(from, "last-week")) {
+    from <- Sys.time() - as.difftime(7, units = "days")
+
+  } else if (identical(from, "last-month")) {
+    from <- Sys.time() - as.difftime(30, units = "days")
+
+  } else if (is.numeric(from)) {
+    from <- Sys.time() - as.difftime(from, units = "days")
+
+  } else if (is.character(from)) {
+    from <- as.POSIXct(from)
+
+  } else {
+    stop("Invalid 'from' argument, please see the docs.")
+  }
+
+  param <- c(
+    end_key = paste0('"', format_iso_8601(from), '"'),
+    limit = if (is.finite(last)) last,
+    descending = "true"
+  )
+
+  url <- paste0(
+    "https://crandb.r-pkg.org:6984/cran/_design/internal/_view/new?",
+    paste0(names(param), "=", param, collapse = "&")
+  )
+
+  rsp <- http_get(url)
+  cnt <- rawToChar(rsp$content)
+  Encoding(cnt) <- "UTF-8"
+  rst <- fromJSON(cnt, simplifyVector = FALSE)
+
+  pkgs <- lapply(rst$rows, function(r) r$value$package)
+  dsc <- rectangle_packages(pkgs)
+
+  dpc <- "Date/Publication"
+  if (dpc %in% colnames(dsc)) {
+    dsc <- dsc[, c(dpc, setdiff(colnames(dsc), dpc))]
+  } else {
+    pub <- tibble::tibble(
+      "Date/Publication" = map_chr(rst$rows, "[[", "key")
+    )
+    dsc <- cbind(pub, dsc)
+  }
+
+  dsc
+}
+
 crandb_query <- function(url, error = TRUE, ...) {
 
   rst <- url0 <- paste0(couchdb_uri(), url)